+++ src/main/java/ch/epfl/leb/sass/models/fluorophores/internal/commands/GenerateFluorophoresGrid2D.java
@@ -17,7 +17,7 @@
  */
 package ch.epfl.leb.sass.models.fluorophores.internal.commands;
 
-import ch.epfl.leb.sass.models.fluorophores.internal.DefaultFluorophore;
+import ch.epfl.leb.sass.models.fluorophores.Fluorophore;
 import ch.epfl.leb.sass.models.fluorophores.internal.dynamics.FluorophoreDynamics;
 import ch.epfl.leb.sass.models.components.Camera;
 import ch.epfl.leb.sass.models.psfs.PSFBuilder;
@@ -101,7 +101,7 @@ public final class GenerateFluorophoresGrid2D implements FluorophoreCommand {
      * @return The list of fluorophores.
      */
     @Override
-    public List<DefaultFluorophore> generateFluorophores() {
+    public List<Fluorophore> generateFluorophores() {
         return FluorophoreReceiver.generateFluorophoresGrid2D(
                 this.spacing, 
                 this.camera,