+++ src/main/java/ch/epfl/leb/sass/models/fluorophores/internal/commands/GenerateFluorophoresRandom3D.java
@@ -17,7 +17,7 @@
  */
 package ch.epfl.leb.sass.models.fluorophores.internal.commands;
 
-import ch.epfl.leb.sass.models.fluorophores.internal.DefaultFluorophore;
+import ch.epfl.leb.sass.models.fluorophores.Fluorophore;
 import ch.epfl.leb.sass.models.fluorophores.internal.dynamics.FluorophoreDynamics;
 import ch.epfl.leb.sass.models.components.Camera;
 import ch.epfl.leb.sass.models.psfs.PSFBuilder;
@@ -116,7 +116,7 @@ public final class GenerateFluorophoresRandom3D implements FluorophoreCommand {
      * @return The list of Fluorophores.
      */
     @Override
-    public List<DefaultFluorophore> generateFluorophores() {
+    public List<Fluorophore> generateFluorophores() {
         return FluorophoreReceiver.generateFluorophoresRandom3D(
                 this.numFluors,
                 this.zLow,